requirements.txt
requests
imdbpy
smtplib
email
shutil
difflib
logging
os
re
time